function_app/requirements.txt
pillow
pillow-heif
numpy
azure-functions
requests

